overhauled project structure
