--- Extension/popup.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>popup html</title>
-
-    <!-- adding bootstrap css -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
-
-    <link rel="stylesheet" href='styles.css'>
-</head>
-<body class='container p-2 m-0'>
-
-    <!-- adding chart.js -->
-    <script src="chart.min.js"></script>
-    <script src="chartjs-plugin-datalabels@2.0.js"></script>
-
-    <div>
-        <table class="table table-bordered w-25 bg-light">
-            <tr>
-                <th class='h4 text-center' colspan='2'>
-                    COVID-19 News Classifier
-                </th>
-            </tr>
-            <tr>
-                <td colspan='2'>
-                    Article URL:
-                    <span class='align-middle text-center' id='article-url' style="font-size: 10px;"></span>
-                </td>
-            </tr>
-            <tr>
-                <td rowspan='2'>
-                    <div style="height: 10em; width: 15em;">
-                        <canvas id="donut"></canvas>
-                    </div>
-                </td>
-                <th class="col align-middle text-center h5" style="font-weight: normal;">
-                    <u>Prediction:</u>
-                </td>
-            </tr>
-            <tr>
-                <td class='align-middle text-center h5' style="min-width: 15em;" id='prediction'>
-                </td>
-            </tr>
-        </table>
-    </div>
-    <script src="script.js"></script>
-</body>
-</html>
-
-<!--
-
-transfer everything and then go from there
-
--> set up table for pop-up
--> figure out how to pass in API to pop-up
--> get typewriter
--> get alerts to work
-
-set up alert based on conspiracy rating
-
-next steps:
-- load from API into results
-- (make article URL invisible) and (existing rows invisible) until API results are found + add a loading screen in-between
-- get percentage to work out when hovering on the pie chart
-
-reset contents
-- clear pie-chart
-- clear
-- create html table for no results found
-
-- add per
--->
