--- Extension/styles.css
@@ -1,8 +1,0 @@
-
-* {
-    box-sizing: border-box;
-}
-
-.table td {
-    text-align: center;
-}
