+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "covid19-classifier"
+version = "0.1.0"
+description = ""
+authors = ["Krish Shah <shahkrish2016@gmail.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+pandas = "^2.1.4"
+numpy = "^1.26.2"
+requests = "^2.31.0"
+bs4 = "^0.0.1"
+lxml = "^4.9.4"
+
+
+[tool.poetry.group.dev.dependencies]
+ipykernel = "^6.27.1"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
